split propagation class
