+++ inspectit-ocelot-core/src/main/java/rocks/inspectit/ocelot/core/instrumentation/context/session/SessionIdManager.java
@@ -1,14 +1,13 @@
 package rocks.inspectit.ocelot.core.instrumentation.context.session;
 
 import com.google.common.annotations.VisibleForTesting;
-import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 import rocks.inspectit.ocelot.core.config.InspectitConfigChangedEvent;
 import rocks.inspectit.ocelot.core.config.InspectitEnvironment;
-import rocks.inspectit.ocelot.core.instrumentation.context.ContextPropagation;
+import rocks.inspectit.ocelot.core.instrumentation.context.propagation.ContextPropagation;
 
 import javax.annotation.PostConstruct;
